Move detectgit into seperate gitutil package
